backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
anthropic==0.7.8
stripe==7.4.0
slowapi==0.1.9
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
psutil==5.9.6
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
jinja2==3.1.2
python-multipart==0.0.6
email-validator==2.1.0
cryptography==41.0.7
bcrypt==4.0.1
python-jose[cryptography]==3.3.0